% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanByNativeDistrib.R
\name{cleanByNativeDistrib}
\alias{cleanByNativeDistrib}
\title{Functions to filter occurrences based on a system of geographical units developped by The International Working Group on Taxonomic Databases for Plant Sciences (TDWG) at approximately "country" level and upwards.}
\usage{
cleanByNativeDistrib(
  point_data,
  species_name,
  species_id = NULL,
  initial_level = 2,
  backbone = "wcvp",
  force.output = FALSE,
  sf = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{point_data}{A two-column data.frame of occurrence records coordinates.}

\item{species_name}{A two character string (Genus species) specifying the name of the species to check.}

\item{species_id}{A character string specifying the id of the species in ipni or Kew world checklist database.}

\item{initial_level}{A numeric integer between 1 and 4. The initial unit level of the TWDG at which the data must be checked against. Default initial level is 2.}

\item{force.output}{A logical. Should the initial point dataset be returned in case where all points have been removed by the cleaning ? Default is FALSE.}

\item{sf}{A logical. Should the results be returned as a sf object ? Default is FALSE.}

\item{...}{Additional parameters to be passed to \code{TDWGinfo()}}
}
\value{
A two-column data.frame or a sf object with point coordinates cleaned by the species native geographic range defined by the TDWG and according to specified criteria.
}
\description{
Clean the species occurrence locations not occurring within its native range.
}
\author{
<i.ondo@kew.org>
}
