% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_id.R
\name{get_species_id}
\alias{get_species_id}
\title{Functions to get the species IPNI ID from the Kew checklist.}
\usage{
get_species_id(species_name, status = "native", level = 2,
  exclude = NULL)
}
\arguments{
\item{species_name}{A two character string (Genus species) specifying the name of the species to search for.}

\item{status}{Optional. A character string specifying the type of geographic region where to search for the species. Can be : 'native', 'introduced' or 'both'}

\item{exclude}{Optional. A character string or a vector of names specifying the geographic region(s) to exclude from the search.
Can be : "NORTHERN AMERICA","SOUTHERN AMERICA", "ANTARCTICA", "EUROPE", "AFRICA", "AUSTRALASIA","ASIA-TEMPERATE", "ASIA-TROPICAL" or "PACIFIC"}
}
\description{
Get the IPNI id of a species that have geographic distribution reported.
}
