% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDistribution.R
\name{queryDistribution}
\alias{queryDistribution}
\title{Functions to get the geographic distribution information of a plant species using the Plant Of the World Online (POWO) API.}
\usage{
queryDistribution(
  binomial,
  ipni_id = NULL,
  status = "natives",
  verbose = TRUE,
  use_acc_name = TRUE,
  ...
)
}
\arguments{
\item{binomial}{A two character string (Genus species) specifying the name of the species to search for.}

\item{ipni_id}{A character string specifying the IPNI id of the species}

\item{status}{Optional. A character string specifying the geographic range where to search for the species. Can be : 'natives' or 'introduced'}

\item{use_acc_name}{A logical. Should the species name be considered 'accepted' ? Default is TRUE.}
}
\description{
Get information about the geographic distribution of a plant species.
}
