% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDistribution.R
\name{queryDistributionId}
\alias{queryDistributionId}
\title{Functions to get the IPNI id of a plant species using the Plant Of the World Online (POWO) API.}
\usage{
queryDistributionId(binomial, status = "natives", use_acc_name = TRUE, ...)
}
\arguments{
\item{binomial}{A two character string (Genus species) specifying the name of the species to search for.}

\item{status}{Optional. A character string specifying the geographic range where to search for the species. Can be : 'natives' or 'introduced'}

\item{use_acc_name}{A logical. Should the species name be considered 'accepted' ? Default is TRUE.}
}
\description{
Get the IPNI id of a species with geographic distribution information.
}
