% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRangeCleaner.R
\name{speciesRangeCleaner}
\alias{speciesRangeCleaner}
\title{Functions to filter species based on a system of geographical units developped by The International Working Group on Taxonomic Databases for Plant Sciences (TDWG) at approximately "country" level and upwards.}
\usage{
speciesRangeCleaner(point_data, species_name, species_id = NULL,
  status = "both", point_fraction = 0.5, unit_fraction = NULL,
  range_filling = NULL, grid_resol = 10000, initial_level = 2,
  by_id = FALSE, recursive = FALSE, sp = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{point_data}{A two-column data.frame of occurrence records coordinates.}

\item{species_name}{A two character string (Genus species) specifying the name of the species to check.}

\item{species_id}{A character string specifying the id of the species in ipni or Kew world checklist database.}

\item{status}{A character string specifying the geographic range where to search for the species. Can be : 'native', 'introduced' or 'both'}

\item{point_fraction}{A numeric between 0 and 1. The minimum fraction of occurrence records that must be included within the geographic range defined in \code{status}.}

\item{unit_fraction}{A numeric between 0 and 1. The minimum fraction of geographical units that must include occurrence records.}

\item{range_filling}{A numeric between 0 and 100. The minimum percent of the geographic area covered by the occurrence records distribution.}

\item{grid_resol}{A numeric string specifying the resolution of the raster grid to be used when calculation the percent coverage of geographical units by point distribution.
Must be given in m. Default is 10000 i.e. 10 km}

\item{initial_level}{A numeric integer between 1 and 4. The initial unit level of the TWDG at which the data must be checked against. Default initial level is 2.}

\item{by_id}{A logical. Should the criteria be applied to each geographical units ? Default is FALSE. Not implemented yet.}

\item{recursive}{A logical. Should the criteria be also applied at the subsequent lower units level ? Default is FALSE.}

\item{sp}{A logical. Should the results be returned as a SpatialPointsDataFrame ? Default is TRUE. Ignored if \code{cleanOcc=FALSE}}

\item{...}{Additional parameters to be passed to \code{TDWGinfo()}}
}
\value{
By default the function returns a logical:
        TRUE if the distribution of species occurrences match the \code{status} range required, given the criteria.
        FALSE otherwise.
}
\description{
Filter the species whose the occurrence distribution data don't match expected distribution data defined by the user.
}
\author{
<i.ondo@kew.org>
}
