% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDWG_info.R
\name{TDWGinfo}
\alias{TDWGinfo}
\title{Functions developed to filter/select species/occurrences based on a system of geographical units
developped by The International Working Group on Taxonomic Databases for Plant Sciences (TDWG) at approximately "country" level and upwards.}
\usage{
TDWGinfo(
  point_data,
  species_name,
  species_id = NULL,
  grid_resol = 10000,
  status = "both",
  use_name_matching = FALSE,
  initial_level = 2,
  backbone = "wcvp",
  by_id = FALSE,
  which_skip = 3,
  recursive = FALSE,
  sf = TRUE,
  proj_utm =
    "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  full_data = FALSE,
  force = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{point_data}{A two-column data.frame of occurrence records coordinates.}

\item{species_name}{A two character string (Genus species) specifying the name of the species to check.}

\item{species_id}{A character string specifying the id of the species in ipni or Kew world checklist.}

\item{grid_resol}{A numeric string specifying the resolution of the raster grid to be used when calculating the percent coverage of geographical units by point distribution.
Must be given in m. Default is 10000 i.e. 10 km. Ignored if number 3 is specified in \code{which_skip} argument.}

\item{use_name_matching}{A logical. Should the species name be matched with the Kew checklist (i.e. look for direct correspondence) ? Default is FALSE.}

\item{initial_level}{A numeric integer between 1 and 4. The initial unit level of the TWDG at which the data must be checked against. Default is level 2.}

\item{by_id}{A logical. Should the criteria be applied to each geographical units ? Default is FALSE.}

\item{which_skip}{A numeric value or a vector of length 2 specifying which point distribution parameter(s) should be skipped from computations.
Set to \code{2} to skip the computation of the proportion of tdwg units sampled; \code{3} to skip the computation of the percent coverage of the tdwg units sampled or\code{NULL} to compute all parameters.}

\item{recursive}{A logical. Should the criteria be also applied at the subsequent lower units level ? Default is FALSE.}

\item{sf}{A logical. Should the results be returned as a sf object ? Default is TRUE.}

\item{proj_utm}{A character string specifying the projection of the raster grid to be used when calculating the percent coverage of geographical units by point distribution.}

\item{full_data}{A logical. Should the initial cleaned dataset be returned ? or the cleaned records only ? Default is FALSE i.e. only cleaned records are returned.}

\item{A}{logical. Should the info be re-computed ? Default is FALSE.}
}
\value{
A sf object if \code{sf=TRUE}, a list object otherwise.
}
\description{
Generates a list of parameters that can be used to filter species and/or occurrence locations based on geographic units recording the native range of species
}
\author{
<i.ondo@kew.org>
}
